💥 Replace the GoogleAppFixture with createGoogleFixtures
