--- src/testing/index.ts
@@ -1,1 +1,0 @@
-export { GoogleAppFixture } from './google-app-fixture.js';
